@@ -2,9 +2,12 @@
 
 namespace App\Controller;
 
+use App\Logger\LoggerKeywords;
 use App\Service\TwitchApiWrapper;
 use App\Service\UserService;
-use Padhie\TwitchApiBundle\Service\TwitchApiService;
+use GuzzleHttp\Exception\ClientException;
+use Padhie\TwitchApiBundle\TwitchAuthenticator;
+use Psr\Log\LoggerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -12,58 +15,71 @@ use Symfony\Component\Routing\Annotation\Route;
 
 final class LoginController extends AbstractController
 {
-    private TwitchApiWrapper $twitchApiWrapper;
-    private UserService $userService;
-
-    public function __construct(TwitchApiWrapper $twitchApiWrapper, UserService $userService)
-    {
-        $this->twitchApiWrapper = $twitchApiWrapper;
-        $this->userService = $userService;
+    public function __construct(
+        private readonly TwitchApiWrapper $twitchApiWrapper,
+        private readonly UserService $userService,
+        private readonly LoggerInterface $logger,
+    ) {
     }
 
-    /**
-     * @Route("/login", name="login")
-     */
+    #[Route('/login', name: 'login')]
     public function indexAction(): Response
     {
         return $this->redirect(
-            $this->twitchApiWrapper->getAccessTokenUrl(
+            $this->twitchApiWrapper->getAccessCodeUrl(
                 array_merge(
-                    TwitchApiService::SCOPE_CHANNEL,
+                    TwitchAuthenticator::SCOPE_CHANNEL,
                     TwitchApiWrapper::SCOPE_PUBSUB
                 )
             )
         );
     }
 
-    /**
-     * @Route("/twitch/get_access", name="twitch_access")
-     */
-    public function getAccessAction(): Response
+    #[Route("/twitch/get_access", name: "twitch_access")]
+    public function getAccessAction(Request $request): Response
     {
-        return $this->render('twitch/access.html.twig');
+        $code = $request->query->get('code') ?? '';
+        $response = $this->twitchApiWrapper->getToken($code);
+
+        return $this->handleAccessToken($request, $response->getAccessToken());
     }
 
-    /**
-     * @Route("/twitch/redirect", name="twitch_redirect")
-     */
+    #[Route("/twitch/redirect", name: "twitch_redirect")]
     public function redirectAction(Request $request): Response
     {
-        $oAuth = $request->get('access_token');
+        $oAuth = $request->get('access_token') ?? null;
+
+        return $this->handleAccessToken($request, $oAuth);
+    }
+
+    private function handleAccessToken(Request $request, ?string $oAuth): Response
+    {
         if ($oAuth === null) {
             return $this->redirectToRoute('frontend');
         }
 
-        $this->twitchApiWrapper->checkAndUseRequestOAuth($request);
+        $this->twitchApiWrapper->checkAndUseRequestOAuth($oAuth);
+
+        $validateModel = $this->twitchApiWrapper->validateByOAuth($oAuth);
+
+        try {
+            $twitchUser = $this->twitchApiWrapper->getUserByName($validateModel->getLogin());
+        } catch (ClientException | \RuntimeException $exception) {
+            $this->logger->error('error during login with twitch', [
+                LoggerKeywords::CODE => 1686688033891,
+                LoggerKeywords::EXCEPTION_MESSAGE => $exception->getMessage(),
+                LoggerKeywords::EXCEPTION_CODE => $exception->getCode(),
+            ]);
+            $this->addFlash('error', 'Unexpected error during login. Please contact the administrator.');
+
+            return $this->redirectToRoute('frontend');
+        }
+
+        $user = $this->userService->getOrCreateUserByTwitchUser($twitchUser, $oAuth);
 
         $session = $request->getSession();
-        $validateModel = $this->twitchApiWrapper->validateByOAuth($oAuth);
-        $session->set(TwitchApiWrapper::SESSION_LOGIN, $validateModel->getLogin());
-
-        $twitchUser = $this->twitchApiWrapper->getUserByName($validateModel->getLogin());
-        $this->userService->getOrCreateUserByTwitchUser($twitchUser, $oAuth);
+        $session->set(UserService::SESSION_KEY_USER_ID, $user->getId());
 
         return $this->redirectToRoute('backend');
     }
-
 }
