@@ -20,9 +20,7 @@ final class ViewController extends AbstractController
         $this->userRepository = $userRepository;
     }
 
-    /**
-     * @Route("/view/{userId}/", name="view")
-     */
+    #[Route("/view/{userId}/", name: "view")]
     public function index(string $userId, Request $request): Response
     {
         $user = $this->userRepository->findOneBy([
@@ -40,7 +38,7 @@ final class ViewController extends AbstractController
         return $this->render('view/index.html.twig', [
             'alerts' => $alerts,
             'user' => $user,
-            'baseFileUrl' => '/sound/' . $user->getId(),
+            'baseFileUrl' => '/sound/' . $user->getId() ?? 0,
             'debug' => (bool) $request->get('debug', false),
         ]);
     }
